requirements.txt
Appium-Python-Client
pytest
allure-pytest
pytest-xdist        # optional, for parallel tests
